requirements.txt
boto3==1.20.18
botocore==1.23.18
greenlet==1.1.2
jmespath==0.10.0
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.2
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
